+++ xill-api/src/main/java/nl/xillio/xill/api/XillEnvironment.java
@@ -18,6 +18,7 @@ package nl.xillio.xill.api;
 import com.google.inject.Injector;
 import nl.xillio.plugins.PluginLoadFailure;
 import nl.xillio.plugins.XillPlugin;
+import nl.xillio.xill.api.components.RobotID;
 import nl.xillio.xill.services.ProgressTracker;
 
 import java.io.IOException;
@@ -69,7 +70,7 @@ public interface XillEnvironment extends AutoCloseable {
 
     /**
      * Sets the {@link XillThreadFactory} used to create threads in plugins.
-     *
+     * <p>
      * Needs to be called before {@link #loadPlugins()}.
      *
      * @param xillThreadFactory the factory
@@ -88,25 +89,54 @@ public interface XillEnvironment extends AutoCloseable {
 
     /**
      * Builds a processor for a specific execution.
+     * <p>
+     * To resolve robots a resource loader is created. This loader builds a chain of paths (directories and archives)
+     * that are traversed when searching for robots. If a Path in this chain is recognised to be a file, it is
+     * interpreted as an archive. Otherwise it is interpreted as a directory.
      *
-     * @param projectRoot the root folder of the project
-     * @param robotPath   the path to the main robot
+     * @param workingDirectory   the working directory of the project
+     * @param fullyQualifiedName the fully qualified name of the robot that should be started
+     * @param robotPath          the robot path
      * @return the processor
      * @throws IOException if we could not load the processor
      */
-    XillProcessor buildProcessor(Path projectRoot, Path robotPath) throws IOException;
+    XillProcessor buildProcessor(Path workingDirectory, String fullyQualifiedName, Path... robotPath) throws IOException;
+
+    /**
+     * Builds a processor for a specific execution.
+     *
+     * @param workingDirectory the working directory of the project
+     * @param robotID          the robot that should be started
+     * @return the processor
+     * @throws IOException if we could not load the processor
+     */
+    XillProcessor buildProcessor(Path workingDirectory, RobotID robotID, Path... robotPath) throws IOException;
 
     /**
      * Builds a processor for a specific execution, with a debugger.
+     * <p>
+     * To resolve robots a resource loader is created. This loader builds a chain of paths (directories and archives)
+     * that are traversed when searching for robots. If a Path in this chain is recognised to be a file, it is
+     * interpreted as an archive. Otherwise it is interpreted as a directory.
      *
-     * @param projectRoot the root folder of the project
-     * @param robotPath   the path to the main robot
-     * @param debugger    the debugger that should be used
+     * @param workingDirectory   the working directory of the project
+     * @param fullyQualifiedName the fully qualified name of the robot that should be started
+     * @param debugger           the debugger that should be used
      * @return the processor
      * @throws IOException if we could not load the processor
      */
-    XillProcessor buildProcessor(Path projectRoot, Path robotPath, Debugger debugger) throws IOException;
+    XillProcessor buildProcessor(Path workingDirectory, String fullyQualifiedName, Debugger debugger, Path... robotPath) throws IOException;
 
+    /**
+     * Builds a processor for a specific execution.
+     *
+     * @param workingDirectory the working directory of the project
+     * @param robotID          the robot that should be started
+     * @param debugger         the debugger that should be used
+     * @return the processor
+     * @throws IOException if we could not load the processor
+     */
+    XillProcessor buildProcessor(Path workingDirectory, RobotID robotID, Debugger debugger, Path... robotPath) throws IOException;
 
     /**
      * Gets a list of all loaded plugins.