+++ xill-api/src/main/java/nl/xillio/xill/api/components/RobotID.java
@@ -15,72 +15,164 @@
  */
 package nl.xillio.xill.api.components;
 
-import java.io.File;
+import nl.xillio.xill.api.XillEnvironment;
+
 import java.io.Serializable;
-import java.util.Hashtable;
-import java.util.Map;
+import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URL;
+import java.net.URLEncoder;
+import java.nio.charset.Charset;
+import java.util.Objects;
+
+import static org.apache.commons.lang3.Validate.notNull;
 
 /**
  * A unique identifier for robots.
+ * <p>
+ * A RobotID is composed of two components: The robot url and a resource uri.
+ * <p>
+ * The <b>robot url</b>, also called <code>url</code>, is an absolute url to the robot resource such that the resource
+ * can be opened and read to get the robot source code.
+ * <p>
+ * The <b>resource uri</b>, or <code>resourceUri</code>, is represented by an {@link URI} relative to the resource root.
+ * This resource root is the root of the parent container of resources. Examples of this could be an archive, folder or
+ * network share.
+ * <p>
+ * The example below describes how these options are defined for a project on the hard drive at "/my/project":
+ * <p>
+ * <code>
+ * // Create the robot url
+ * URL robotUrl = new URL("file:///my/project/libs/URLParser.xill");
+ * <p>
+ * // Create the resourceUri relative to the project root
+ * URI resourceUri = URI.create("libs/URLParser.xill");
+ * <p>
+ * // Create the RobotID
+ * RobotID robotId = new RobotID(robotUrl, resourceUri);
+ * </code>
  */
 public class RobotID implements Serializable {
-    private static Map<String, RobotID> ids = new Hashtable<>();
-    private final File path;
-    private final File projectPath;
+    private final URL url;
+    private final URI resourceUri;
 
-    private RobotID(final File path, final File projectPath) {
-        this.path = path;
-        this.projectPath = projectPath;
+    /**
+     * Create a new robotID from an url and a resource path.
+     *
+     * @param url          The url to this robot
+     * @param resourcePath The robot path relative to the resource root it will be escaped using percentage encoding
+     * @throws IllegalArgumentException if resourcePath is not a valid relative path
+     */
+    public RobotID(URL url, String resourcePath) {
+        this(url, escape(resourcePath));
     }
 
     /**
-     * Returns the path associated with this id.
+     * Create a new robotID from an url and a resource uri.
      *
-     * @return the path associated with this id
+     * @param url         The url to this robot
+     * @param resourceUri The robot uri relative to the resource root
+     * @throws IllegalArgumentException if resourceUri is not a relative uri
      */
-    public File getPath() {
-        return path;
+    public RobotID(URL url, URI resourceUri) {
+        notNull(url);
+        notNull(resourceUri);
+
+        if (resourceUri.isAbsolute() ||
+                resourceUri.getScheme() != null ||
+                resourceUri.getAuthority() != null) {
+            throw new IllegalArgumentException(resourceUri + " is not a relative uri");
+        }
+
+        this.url = url;
+        this.resourceUri = resourceUri;
+    }
+
+    public URL getURL() {
+        return url;
+    }
+
+    public URI getResourceUri() {
+        return resourceUri;
+    }
+
+    public String getResourcePath() {
+        return resourceUri.getPath();
     }
 
     @Override
     public String toString() {
-        return path.getAbsolutePath();
+        return "RobotID{" +
+                "url=" + url +
+                ", resourceUri=" + resourceUri +
+                '}';
     }
 
-    /**
-     * Gets/creates a robotID that is singular for every path.
-     *
-     * @param file        the robot file
-     * @param projectPath the path to the root folder of the workspace
-     * @return a unique robot id for this path
-     */
-    public static RobotID getInstance(final File file, final File projectPath) {
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        RobotID robotID = (RobotID) o;
+        return Objects.equals(url, robotID.url) &&
+                Objects.equals(getResourcePath(), robotID.getResourcePath());
+    }
 
-        String identity = file.getAbsolutePath() + "in" + projectPath.getAbsolutePath();
+    @Override
+    public int hashCode() {
+        return Objects.hash(url, resourceUri);
+    }
 
-        RobotID id = ids.get(identity);
-
-        if (id == null) {
-            id = new RobotID(file, projectPath);
-            ids.put(identity, id);
+    private static URL createUrl(URL baseUrl, String resourcePath) {
+        try {
+            return new URL(baseUrl, resourcePath);
+        } catch (MalformedURLException e) {
+            throw new IllegalArgumentException(e);
         }
-
-        return id;
     }
 
     /**
-     * @return the projectPath
-     */
-    public File getProjectPath() {
-        return projectPath;
-    }
-
-    /**
-     * Used in tests to create a dummy ID.
+     * Build a RobotID from a baseUrl and a resourcePath.
      *
-     * @return a dummy IDfor testing.
+     * @param baseUrl      The url to the resource root that contains this robot
+     * @param resourcePath The robot uri relative to the baseUrl
+     * @return the RobotID
+     */
+    public static RobotID build(String baseUrl, String resourcePath) {
+        return new RobotID(createUrl(createUrl(null, baseUrl), resourcePath), escape(resourcePath));
+    }
+
+    /**
+     * Build a dummy ID that points to an arbitrary robot.
+     *
+     * @return a dummy robotID
      */
     public static RobotID dummyRobot() {
-        return new RobotID(new File("."), new File("."));
+        return build("file:///path/to/project", "my/Robot.xill");
+    }
+
+
+    private static URI escape(String resourcePath) {
+        String[] parts = resourcePath.split("[\\\\|/]");
+
+        for (int i = 0; i < parts.length; i++) {
+            try {
+                parts[i] = URLEncoder.encode(parts[i], Charset.defaultCharset().name())
+                        .replace("+", "%20");
+            } catch (UnsupportedEncodingException e) {
+                throw new IllegalStateException("The default character set could not be found");
+            }
+        }
+        return URI.create(String.join("/", parts));
+    }
+
+    /**
+     * Convert a qualified name to a path.
+     *
+     * @param fullyQualifiedName the qualified name to convert to a path
+     * @return the converted qualified name to a path
+     */
+    public static String qualifiedNameToPath(String fullyQualifiedName) {
+        return fullyQualifiedName.replace('.', '/') + XillEnvironment.ROBOT_EXTENSION;
     }
 }